Add files via upload
This is the very first round of code I had implemented to gain a simple understanding of the data I am dealing with. And what best way than to drop a correlation between the features.
Unfortunately, even though there is a slight linear relation between wind speed and wave height it is not good enough for prediction. 
As shown in the implementation the score is very less.

PS: If you have a better approach or a different model that works wonderfully with this type of data, do share it with me.